# run `register_llvm_rules` in a module where optimizations are disabled