@vchuravy 2024-02-23
I still need to confirm, but iirc this should only apply to the ccallable thunk.