@vchuravy 2026-09-05
Superseded by #3531.

Measured this on today's tree: as written this PR now recovers 5s of the 22s available. The mechanism moved out from under it. In 2023 `@augfunc` built an anonymous closure *inside* the low-optimization module, so `@compiler_options` covered the actual `@cfunction` target. Today `@register_aug`/`_rev`/`_fwd`/`_diffuse` define named `<rule>_cfunc` shims in `Compiler`, which stay fully optimized and `@inline` the rule bodies, so moving only the `@cfunction` call site leaves all 115 rule bodies in codegen.

#3531 moves the shim definitions as well: precompile goes 47.9s -> 37.8s as shipped, or 33.3s -> 11.1s with the `@compile_workload` disabled.

On the concern about adjoint generation runtime -- it is real but small. A deliberately generic-dispatch-heavy differentiation goes 5.18s -> 5.42s; scalar and array gradients do not move, and neither does running an already-compiled gradient.